# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: support-management
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "12.4"
servers:
  - url: http://localhost:51578
    description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/metadata/external-id-types:
    get:
      tags:
        - Metadata for external ID types
      summary: Get externalIdTypes
      description: Get all externalIdTypes for the namespace and municipality
      operationId: getExternalIdTypes
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExternalIdType"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - Metadata for external ID types
      summary: Create externalIdType
      description: Create new externalIdType for the namespace and municipality
      operationId: createExternalIdType
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalIdType"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/metadata:
    get:
      tags:
        - Metadata
      summary: Get all metadata
      description: Get all metadata for provided namespace and municipality
      operationId: getAll
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/metadata/external-id-types/{externalIdType}:
    get:
      tags:
        - Metadata for external ID types
      summary: Get externalIdType
      description: "Get externalIdType matching sent in namespace, municipality and\
        \ externalIdType"
      operationId: getExternalIdType
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: externalIdType
          in: path
          description: Name of external ID type
          required: true
          schema:
            type: string
          example: PRIVATE
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalIdType"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Metadata for external ID types
      summary: Delete externalIdType
      description: "Delete external ID type matching namespace, municipality and externalIdType"
      operationId: deleteExternalIdType
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: externalIdType
          in: path
          description: Name of externalIdType
          required: true
          schema:
            type: string
          example: PRIVATE
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: { }
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Label:
      type: object
      description: Label model
      properties:
        id:
          type: string
          description: Label ID
          examples:
            - 5f79a808-0ef3-4985-99b9-b12f23e202a7
        classification:
          type: string
          description: Label classification
          examples:
            - subtype
          minLength: 1
        displayName:
          type: string
          description: Display name for the label
          examples:
            - Nyckelkort
        resourcePath:
          type: string
          description: Resource path
          examples:
            - /parent/child/keycard
          readOnly: true
        resourceName:
          type: string
          description: Resource name
          examples:
            - keycard
          minLength: 1
          pattern: "[A-Z0-9_]+"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
      required:
        - classification
        - resourceName
...
    Status:
      type: object
      description: Status model
      properties:
        name:
          type: string
          description: Name for the status
          examples:
            - statusName
          minLength: 1
        created:
          type: string
          format: date-time
          description: Timestamp when the status was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the status was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
      required:
        - name
    Role:
      type: object
      description: Role model
      properties:
        name:
          type: string
          description: Name for the role. Used as key
          examples:
            - roleName
          minLength: 1
        displayName:
          type:
            - string
            - "null"
          description: Display name for the role
          examples:
            - Role name
        created:
          type: string
          format: date-time
          description: Timestamp when the role was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the role was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
      required:
        - name
    ExternalIdType:
      type: object
      description: ExternalIdType model
      properties:
        name:
          type: string
          description: Name for the external id type
          examples:
            - PRIVATE
          minLength: 1
        created:
          type: string
          format: date-time
          description: Timestamp when the external id type was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the external id type was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
      required:
        - name
    ContactReason:
      type: object
      description: Contact reason model
      properties:
        id:
          type: integer
          format: int64
          description: ID
          examples:
            - "123"
          readOnly: true
        reason:
          type: string
          description: Reason for contact
          examples:
            - Segt internet
          minLength: 1
        created:
          type: string
          format: date-time
          description: Timestamp when the contact reason was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the contact reason was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
      required:
        - reason
    Category:
      type: object
      description: Category model
      properties:
        name:
          type: string
          description: Name for the category
          examples:
            - Category name
        displayName:
          type: string
          description: Display name for the category
          examples:
            - Displayed name
        types:
          type: array
          items:
            $ref: "#/components/schemas/Type"
          uniqueItems: true
        created:
          type: string
          format: date-time
          description: Timestamp when the category was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the category was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
    Type:
      type: object
      description: Type model
      properties:
        name:
          type: string
          description: Name for the type
          examples:
            - typename
          minLength: 1
        displayName:
          type: string
          description: Display name for the type
          examples:
            - Displayed name
        escalationEmail:
          type: string
          format: email
          description: Email for where to escalate the errand if needed
          examples:
            - escalationgroup@sesamestreet.com
        created:
          type: string
          format: date-time
          description: Timestamp when type was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when type was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
      required:
        - name
...
    Labels:
      type: object
      description: Labels model
      properties:
        labelStructure:
          type: array
          items:
            $ref: "#/components/schemas/Label"
      readOnly: true
    MetadataResponse:
      type: object
      description: MetadataResponse model
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        externalIdTypes:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdType"
        labels:
          $ref: "#/components/schemas/Labels"
          readOnly: true
        statuses:
          type: array
          items:
            $ref: "#/components/schemas/Status"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        contactReasons:
          type: array
          items:
            $ref: "#/components/schemas/ContactReason"
...
  securitySchemes: { }